SCKAN-275 refactor: Split ingestion script in multiple modules
