+++ backend/composer/services/cs_ingestion/helpers/validators.py
@@ -1,0 +1,91 @@
+from typing import List, Dict, Any, Set
+
+from neurondm import orders
+
+from composer.models import ConnectivityStatement, Sex, Specie
+from composer.services.cs_ingestion.helpers.common_helpers import ID, VALIDATION_ERRORS, ORIGINS, DESTINATIONS, VIAS, \
+    found_entity, \
+    SEX, SPECIES, FORWARD_CONNECTION
+from composer.services.cs_ingestion.logging_service import LoggerService
+from composer.services.cs_ingestion.models import ValidationErrors, LoggableAnomaly
+
+logger_service = LoggerService()
+
+
+def validate_statements(statements: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
+    db_reference_uris = set(ConnectivityStatement.objects.values_list('reference_uri', flat=True))
+    input_statement_ids = {statement[ID] for statement in statements}
+    statement_ids = input_statement_ids.union(db_reference_uris)
+
+    for statement in statements:
+        # Initialize validation_errors if not already present
+        if VALIDATION_ERRORS not in statement:
+            statement[VALIDATION_ERRORS] = ValidationErrors()
+
+        # Validate entities, sex, and species, updating validation_errors accordingly
+        annotate_invalid_entities(statement)
+        annotate_invalid_sex(statement)
+        annotate_invalid_species(statement)
+
+        # Validate forward connection
+        annotate_invalid_forward_connections(statement, statement_ids)
+
+    return statements
+
+
+def annotate_invalid_entities(statement: Dict) -> bool:
+    # TODO: Update validators based on updateAnatomicalEntities flag
+    has_invalid_entities = False
+
+    entities_to_check = list(statement[ORIGINS].anatomical_entities)
+    entities_to_check.extend(entity for dest in statement[DESTINATIONS] for entity in dest.anatomical_entities)
+    entities_to_check.extend(entity for via in statement[VIAS] for entity in via.anatomical_entities)
+
+    for entity in entities_to_check:
+        if isinstance(entity, orders.rl):
+            region_found = found_entity(entity.region)
+            layer_found = found_entity(entity.layer)
+            if not region_found:
+                statement[VALIDATION_ERRORS].entities.add(entity.region)
+                has_invalid_entities = True
+            if not layer_found:
+                statement[VALIDATION_ERRORS].entities.add(entity.layer)
+                has_invalid_entities = True
+        else:
+            uri = str(entity)
+            if not found_entity(uri):
+                statement[VALIDATION_ERRORS].entities.add(uri)
+                has_invalid_entities = True
+
+    return has_invalid_entities
+
+
+def annotate_invalid_sex(statement: Dict) -> bool:
+    if statement[SEX]:
+        if len(statement[SEX]) > 1:
+            logger_service.add_anomaly(
+                LoggableAnomaly(statement[ID], None, f'Multiple sexes found in statement.'))
+
+            first_sex_uri = statement[SEX][0]
+            if not Sex.objects.filter(ontology_uri=first_sex_uri).exists():
+                statement[VALIDATION_ERRORS].sex.add(first_sex_uri)
+            return True
+    return False
+
+
+def annotate_invalid_species(statement: Dict) -> bool:
+    has_invalid_species = False
+    for species_uri in statement[SPECIES]:
+        if not Specie.objects.filter(ontology_uri=species_uri).exists():
+            statement[VALIDATION_ERRORS].species.add(species_uri)
+            has_invalid_species = True
+    return has_invalid_species
+
+
+def annotate_invalid_forward_connections(statement: Dict, statement_ids: Set[str]) -> bool:
+    has_invalid_forward_connection = False
+    for reference_uri in statement[FORWARD_CONNECTION]:
+        if reference_uri not in statement_ids:
+            statement[VALIDATION_ERRORS].forward_connection.add(reference_uri)
+            has_invalid_forward_connection = True
+    return has_invalid_forward_connection
